+++ tests/test_glm_asr_nano.rs
@@ -22,7 +22,7 @@ fn glm_asr_nano_generate() -> Result<()> {
                         "type": "audio",
                         "audio_url": 
                         {
-                            "url": "file://./assets/audio/zh.mp3"
+                            "url": "file://./assets/audio/voice_01.wav"
                         }
                     },           
                     {