+++ docs/changelog.md
@@ -5,6 +5,12 @@ All notable changes to aha will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### 2026-04-17
+- Qwen3ASR add vad data recognition
+
+### 2026-04-16
+- fix FireRedVAD fsmn cache bug
+
 ### 2026-04-15
 - add FireRedVAD